refactor: refactor boards module

+++ src/resources/boards/board.service.js
@@ -2,13 +2,16 @@ const boardsRepo = require('./board.memory.repository');
 const tasksRepo = require('../tasks/task.service');
 const Board = require('./board.model');
 
-const getAllBoards = () => boardsRepo.getAllBoards();
-const createBoard = (board) => (Board.validateBoard(board)) ? boardsRepo.createBoard(new Board(board)) : false;
-const updateBoard = (boardId, updatedBoardData) => (Board.validateBoard(updatedBoardData)) ? boardsRepo.updateBoard(boardId, updatedBoardData) : false;
-const deleteBoard = (boardId) => {
-    const boardTasks = tasksRepo.getTasksByBoardId(boardId);
-    boardTasks.map((task) => tasksRepo.deleteTask(boardId, task.id))
-    return boardsRepo.deleteBoard(boardId);
+const getAllBoards = async () => boardsRepo.getAllBoards();
+const createBoard = async (board) => (Board.validateBoard(board)) ? boardsRepo.createBoard(new Board(board)) : false;
+const updateBoard = async (boardId, updatedBoardData) => boardsRepo.updateBoard(boardId, updatedBoardData);
+const deleteBoard = async (boardId) => {
+    const deletedBoard = boardsRepo.deleteBoard(boardId);
+    if(deletedBoard) {
+        const boardTasks = tasksRepo.getTasksByBoardId(boardId);
+        boardTasks.map((task) => tasksRepo.deleteTask(boardId, task.id));
+    }
+    return deletedBoard;
 }
-const getBoardById = (boardId) => boardsRepo.getBoardById(boardId)
+const getBoardById = async (boardId) => boardsRepo.getBoardById(boardId)
 module.exports = { getAllBoards, createBoard, getBoardById, updateBoard, deleteBoard };
